Multiple minor optimizations and bug fixes
Reworked internal file system to be more organized and easier for future use.

Fixed the bot being able to used as a plugin on a minecraft server!
- You can now just throw this bot in your plugins folder and it will work on its own!!!
- This will open up future commands to be used in game, and the ability to access server information from your discord server.

Added a Plugin Mode: yes/no for the version to command to indicate if the bot is being used as a plugin, or standalone.

Reworked a ton of internal organization