Added .profile [user] command, this will be more used in the future, just getting it setup for now. Added .role_all <role> command, to give that role to all users.
Implemented a new queueing system for giving roles. This will be used for any event or command that gives a user a role. This is to help out welcome roles being given if a mass amount of members join at the same time. Discords API limits how fast you can give roles, so this is a fix for that.

Added TimeUtil.java